qualifiers/whistleblower/src/requirements.txt
flask
pycrypto
